+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Replace sftp `mockClosableClient` type with mockery.
 
 ## [[v7.20.1](https://github.com/C2FO/vfs/releases/tag/v7.20.1)] - 2026-06-22
 ### Documentation